feat(sdk-core): move ntilde (de/se)rialize methods
TICKET: BG-75009
